% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popavg_dist.R
\name{popavg_dist}
\alias{popavg_dist}
\title{Population average distribution function}
\usage{
popavg_dist(
  x,
  knots,
  logk0 = -0.1,
  g0 = 0.1,
  delta_vec,
  h_parm = 1,
  frailty = "PS",
  xi = 0.99,
  n = 0.99,
  s = 1.01
)
}
\arguments{
\item{x}{The domain of the function; the survival times.}

\item{knots}{A vector of the internal knots or change-points. Must be within
the range (min(x), max(x)).}

\item{logk0}{The log of the shape parameter for the first piece.}

\item{g0}{The shape parameter.}

\item{delta_vec}{the vector of delta parameters, in order.}

\item{h_parm}{The heterogeneity parameter or homogeneity parameter.  You
decide what h stands for.}

\item{frailty}{character matching "PS" for positive stable (alpha case in
paper), "GA" for gamma (beta),
"IG" for Inverse Gaussian (lambda), "TP1" (rho) for two point with mean of 1,
and "TPU" for two point unrestricted (omega).}

\item{xi}{a scalar less than 1 for the \code{frailty="TP1"} case.}

\item{n}{a scalar less than 1 for the \code{frailty="TPU"} case.}

\item{s}{a scalar more than 1 for the \code{frailty="TPU"} case.}
}
\value{
Value
}
\description{
Population average distribution function
}
\examples{
rnorm(1)
}
