% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjspec_haz.R
\name{subjspec_haz}
\alias{subjspec_haz}
\title{Subject specific hazard function}
\usage{
subjspec_haz(x, knots, logk0 = -0.1, g0 = 0.1, delta_vec)
}
\arguments{
\item{x}{The domain of the function; the survival times.}

\item{knots}{A vector of the internal knots or change-points. Must be within
the range (min(x), max(x)).}

\item{logk0}{The log of the shape parameter for the first piece.}

\item{g0}{The shape parameter.}

\item{delta_vec}{the vector of delta parameters, in order.}
}
\value{
Value
}
\description{
Take the parameters estimated from \code{popavg_dist} and calculate
subject specific hazards of the Weibull form \code{b*k*x^(k-1)}.
}
\examples{
rnorm(1)
}
